Revert "exercise-03"
This reverts commit d9f5cde6ce6173de8c3468debf1551de333595a3.
