--- src/ErrorPage.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-
-const ErrorPage = () =>
-
-    <div className="App">
-        
-            <div>
-                <h1>Error 404</h1>
-                <p>This page does not exist.</p>
-            </div>
-        
-    </div>
-
-export default ErrorPage;
